refactor: moved image validation functions and corresponding tests for improved input handling
