+++ tests/entities/images/test_entities_images_validators.py
@@ -1,15 +1,12 @@
 import numpy as np
 import pytest
 from PIL import Image as PILImage
-from sklearn.base import BaseEstimator
 
 from siapy.core.exceptions import (
     InvalidInputError,
     InvalidTypeError,
-    MethodNotImplementedError,
 )
-from siapy.utils.validators import (
-    check_model_prediction_methods,
+from siapy.entities.images.validators import (
     validate_image_size,
     validate_image_to_numpy,
     validate_image_to_numpy_3channels,
@@ -80,33 +77,3 @@ def test_validate_image_size():
 
     with pytest.raises(InvalidInputError):
         validate_image_size((100, "150"))
-
-
-class MockModelValid(BaseEstimator):
-    def fit(self, X, y):
-        pass
-
-    def predict(self, X):
-        pass
-
-    def score(self, X, y):
-        pass
-
-
-class MockModelInvalid(BaseEstimator):
-    def fit(self, X, y):
-        pass
-
-
-def test_check_model_prediction_methods_valid():
-    model = MockModelValid()
-    check_model_prediction_methods(model)
-
-
-def test_check_model_prediction_methods_invalid():
-    model = MockModelInvalid()
-    with pytest.raises(
-        MethodNotImplementedError,
-        # match="The model must have methods: 'fit', 'predict', and 'score'.",
-    ):
-        check_model_prediction_methods(model)
